Rename ldk node to bayWalletNode
